+++ core/testdata/testspecs/v2_specs.go
@@ -828,3 +828,34 @@ storeBlockhashesBatchSize = %d
 
 	return BlockHeaderFeederSpec{BlockHeaderFeederSpecParams: params, toml: toml}
 }
+
+type StreamSpecParams struct {
+	Name string
+}
+
+type StreamSpec struct {
+	StreamSpecParams
+	toml string
+}
+
+// Toml returns the BlockhashStoreSpec in TOML string form.
+func (b StreamSpec) Toml() string {
+	return b.toml
+}
+
+func GenerateStreamSpec(params StreamSpecParams) StreamSpec {
+	template := `
+type = "stream"
+schemaVersion = 1
+name = "%s"
+observationSource = """
+ds          [type=http method=GET url="https://chain.link/ETH-USD"];
+ds_parse    [type=jsonparse path="data,price"];
+ds_multiply [type=multiply times=100];
+ds -> ds_parse -> ds_multiply;
+"""
+`
+
+	toml := fmt.Sprintf(template, params.Name)
+	return StreamSpec{StreamSpecParams: params, toml: toml}
+}