+++ integration-tests/smoke/cron_test.go
@@ -25,7 +25,6 @@ func TestCronBasic(t *testing.T) {
 		WithMockAdapter().
 		WithCLNodes(1).
 		WithStandardCleanup().
-		WithLogStream().
 		Build()
 	require.NoError(t, err)
 
@@ -72,7 +71,6 @@ func TestCronJobReplacement(t *testing.T) {
 		WithMockAdapter().
 		WithCLNodes(1).
 		WithStandardCleanup().
-		WithLogStream().
 		Build()
 	require.NoError(t, err)